Chore: Change the common popup structure to use with email popup
