+++ src/app/posts/[slug]/page.tsx
@@ -3,7 +3,6 @@ import Comments from '@/components/Comments';
 import PostContent from '@/components/PostContent';
 import TableOfContents from '@/components/TableOfContents';
 import TopButton from '@/components/ui/TopButton';
-import { PopUpContextProvider } from '@/context/PopUpContext';
 import { getFeaturedPosts, getPostData } from '@/service/posts';
 import { Metadata } from 'next';
 import { notFound } from 'next/navigation';
@@ -30,22 +29,20 @@ export default async function PostPage({ params: { slug } }: Props) {
     const { prevPost, nextPost } = post;
 
     return (
-      <PopUpContextProvider>
-        <article className='relative flex mt-10'>
-          <div className='w-full lg:basis-0 lg:grow-[4] lg:w-4/5'>
-            <PostContent post={post} />
-            <section className='overflow-hidden flex flex-col px-4 md:flex-row'>
-              {prevPost && <AdjacentPostCard type='prev' post={prevPost} />}
-              {nextPost && <AdjacentPostCard type='next' post={nextPost} />}
-            </section>
-            <Comments postId={post.id} />
-          </div>
-          <aside className='max-lg:hidden sticky top-44 left-0 p-2 w-1/5 h-full basis-0 shrink-0 grow'>
-            <TableOfContents content={post.content} />
-          </aside>
-          <TopButton />
-        </article>
-      </PopUpContextProvider>
+      <article className='relative flex mt-10'>
+        <div className='w-full lg:basis-0 lg:grow-[4] lg:w-4/5'>
+          <PostContent post={post} />
+          <section className='overflow-hidden flex flex-col px-4 md:flex-row'>
+            {prevPost && <AdjacentPostCard type='prev' post={prevPost} />}
+            {nextPost && <AdjacentPostCard type='next' post={nextPost} />}
+          </section>
+          <Comments postId={post.id} />
+        </div>
+        <aside className='max-lg:hidden sticky top-44 left-0 p-2 w-1/5 h-full basis-0 shrink-0 grow'>
+          <TableOfContents content={post.content} />
+        </aside>
+        <TopButton />
+      </article>
     );
   } catch (error) {
     notFound();
